added registration auth filtering posts linking posts

+++ controllers/post.controller.js
@@ -1,12 +1,29 @@
 const models = require('../models');
 
 const showPosts = (req, res) => {
+  const { authorId } = req.query;
+  let params = {
+    include: {
+      model: models.User,
+      attributes: [ 'name' ],
+      required: true
+    }
+  };
+  if (authorId) {
+    params = {
+      where: {
+        userId: authorId
+      },
+      ...params
+    };
+  }
   models.Post
-    .findAll()
+    .findAll(params)
     .then((result) => {
       res.status(200).json(result);
     })
     .catch((error) => {
+      console.log(error);
       res.status(500).json({
         message: 'Something went wrong!',
         err: error
@@ -17,7 +34,13 @@ const showPosts = (req, res) => {
 const showPost = (req, res) => {
   const { id } = req.params;
   models.Post
-    .findByPk(id)
+    .findByPk(id, {
+      include: {
+        model: models.User,
+        attributes: [ 'name' ],
+        required: false
+      }
+    })
     .then((result) => {
       if (result) {
         res.status(200).json(result);
@@ -35,9 +58,10 @@ const showPost = (req, res) => {
 };
 
 const createPost = (req, res) => {
+  const { userId } = req.userData;
   const { content } = req.body;
   models.Post
-    .create({ content })
+    .create({ content, userId })
     .then((result) => {
       res.status(201).json({
         massage: 'Post created successfully',
@@ -53,10 +77,11 @@ const createPost = (req, res) => {
 };
 
 const updatePost = (req, res) => {
+  const { userId } = req.userData;
   const { id } = req.params;
   const { updatedPost } = req.body;
   models.Post
-    .update({ content: updatedPost }, { where: { id } })
+    .update({ content: updatedPost }, { where: { id, userId } })
     .then((results) => {
       if (!!results[0]) {
         res.status(200).json({
@@ -78,13 +103,20 @@ const updatePost = (req, res) => {
 };
 
 const deletePost = (req, res) => {
+  const { userId } = req.userData;
   const { id } = req.params;
   models.Post
-    .destroy({ where: { id } })
-    .then(() => {
-      res.status(200).json({
-        message: 'Post deleted successfully'
-      });
+    .destroy({ where: { id, userId } })
+    .then((result) => {
+      if (result === 1) {
+        res.status(200).json({
+          message: 'Post deleted successfully'
+        });
+      } else {
+        res.status(404).json({
+          message: 'Post not found!'
+        });
+      }
     })
     .catch((error) => {
       res.status(500).json({
